--- src/data/src/main/java/com/gabrielbmoro/moviedb/repository/mappers/FavoriteMovieMapper.kt
@@ -1,22 +1,0 @@
-package com.gabrielbmoro.moviedb.repository.mappers
-
-import com.gabrielbmoro.moviedb.domain.entities.Movie
-import com.gabrielbmoro.moviedb.repository.datasources.room.dto.FavoriteMovieDTO
-
-class FavoriteMovieMapper {
-
-    fun map(id: Int? = null, movie: Movie): FavoriteMovieDTO {
-        return FavoriteMovieDTO(
-            id = id,
-            votesAverage = movie.votesAverage,
-            title = movie.title,
-            language = movie.language,
-            overview = movie.overview,
-            posterImageUrl = movie.posterImageUrl,
-            backdropImageUrl = movie.backdropImageUrl,
-            popularity = movie.popularity,
-            releaseDate = movie.releaseDate,
-            movieId = movie.id
-        )
-    }
-}
